Using one handler for both registrations
